Move reducers and actions into seperate folders
