+++ .eslintrc.js
@@ -4,7 +4,4 @@ module.exports = {
     "browser": true,
     "node": true,
   },
-  "rules": {
-    "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }]
-  },
 };
