+++ src/client/react/actions/search.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/prefer-default-export
 export const type = typedValue => ({
-  type: 'TYPE',
+  type: 'SEARCH/TYPE',
   typedValue,
 });
