+++ src/client/react/reducers/search.js
@@ -1,0 +1,21 @@
+const DEFAULT_STATE = {
+  searchInput: '',
+  searchResults: [],
+};
+
+const search = (state = DEFAULT_STATE, action) => {
+  switch (action.type) {
+    case 'SEARCH/TYPE':
+      return {
+        ...state,
+        searchInput: action.typedValue,
+        searchResults: [
+          { type: 's', name: '18561' },
+        ],
+      };
+    default:
+      return state;
+  }
+};
+
+export default search;
