+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/client/react/index.js',
+  entry: './src/client/react/index.jsx',
   output: {
     path: path.resolve(__dirname, 'src/client/static'),
     filename: 'bundle.js',
@@ -9,11 +9,14 @@ module.exports = {
   module: {
     rules: [
       {
-        test: [/\.js$/],
+        test: [/\.js$/, /\.jsx$/],
         exclude: [/node_modules/],
         loader: 'babel-loader',
         options: { presets: ['es2015', 'react', 'stage-2'] },
       },
     ],
   },
+  resolve: {
+    extensions: ['.js', '.jsx'],
+  }
 };
